TST: move grouping test to correct location
